Adding DAC diagnostics
Fixing handle size on retina
Improving performance of ScrollableView (canvas and 3D visualiser)

+++ src/dacs/LaserDock/ofxLaserDacLaserDock.cpp
@@ -152,7 +152,7 @@ void DacLaserDock :: threadedFunction(){
                     }
                     unlock();
                     if(dacDevice->set_dac_rate(newPPS)) {
-                        pps = newPPS;
+                        pps = (uint32_t)newPPS;
                     } else {
                         // ?? do something?
                     }
